+++ CHANGELOG.md
@@ -11,3 +11,17 @@
     * `highest`: Instance method that returns the highest value of `this` or `other`.
     * `lowest`: Instance method that returns the lowest value of `this` or `other`.
     * `inBetween`: Instance method that checks if `this` is between two specified values.
+
+## 0.0.1-alpha.3
+
+### Notes:
+  * Dropped dependency `intl` down to `^0.18.0` (from `^0.18.1`) to avoid conflicts.
+  
+### Additions:
+
+#### Extensions:
+  * `IterableStringPrime`:
+    * `whereNotEmpty`: Instance method that returns all non-empty values in the source collection.
+
+  * `IterableOptionalStringPrime`:
+    * `whereNotEmpty`: Instance method that returns all non-empty values in the source collection.